refactor: move fixtures and helpers out of test_evalhub_job_logs_mt.py
Move 3 fixtures to multitenancy/conftest.py and 3 helper functions to
utils.py to follow repo convention of keeping test files free of
fixture and utility definitions.
